adapt ubuntu14.04 x86_64

+++ ubuntu/ubuntu14.04/qcow2/x86_64/build-vm.sh
@@ -1,38 +1,63 @@
 #!/bin/bash
 
-virsh destroy kdev-ubuntu14 || :
-virsh undefine kdev-ubuntu14 --nvram || :
+set -x
 
-if [ -f rootfs.qcow2 ];then
+ISOURL="https://old-releases.ubuntu.com/releases/14.04.4/ubuntu-14.04-server-amd64.iso"
+FILE_SERVER_PORT="63333"
+VMNAME="kdev-ubuntu14"
+ISONAME=$(basename ${ISOURL})
+
+virsh destroy ${VMNAME} || :
+virsh undefine ${VMNAME} --nvram || :
+
+if [ -f rootfs.qcow2 ]; then
+	lsof rootfs.qcow2
+	if [ $? -eq 0 ]; then
+		echo "rootfs.qcow2 is inuse"
+		exit 1
+	fi
 	rm -f rootfs.qcow2
-fi
-
-if [ ! -f rootfs.qcow2 ];then
+	qemu-img create -f qcow2 rootfs.qcow2 500G
+else
 	qemu-img create -f qcow2 rootfs.qcow2 500G
 fi
 
 sync
-sleep 1
 
-if [ ! -f ubuntu-14.04-server-amd64.iso ]; then
-	wget -c https://old-releases.ubuntu.com/releases/14.04.4/ubuntu-14.04-server-amd64.iso
+if [ ! -f "${ISONAME}" ]; then
+	wget -c ${ISOURL}
 fi
 
-virt-install --connect qemu:///system \
-	--name kdev-ubuntu14 \
-	--os-variant ubuntu14.04 \
-	--ram 4096 \
-	--vcpus 16 \
-	--graphics none \
-	--disk path=rootfs.qcow2,size=8,format=qcow2,bus=scsi,target.dev=sda \
-	--controller type=scsi,model=auto \
-	--location ubuntu-14.04-server-amd64.iso \
+virt-install \
+	--name ${VMNAME} \
+	--arch x86_64 \
+	--vcpus 4 \
+	--ram 2048 \
+	--boot cdrom,hd \
+	--network network=default,model=e1000e \
+	--disk path=$(pwd)/rootfs.qcow2,format=qcow2,bus=sata \
+	--location $(pwd)/${ISONAME},initrd=install/initrd.gz,kernel=install/vmlinuz \
 	--initrd-inject preseed.cfg \
-	--extra-args="auto=true priority=critical preseed/file=/preseed.cfg console=tty0 console=ttyS0,115200 autoinstall" \
-	--check all=off
+	--extra-args="auto=true priority=critical preseed/file=/preseed.cfg earlyprintk console=tty0 console=ttyS0,115200 " \
+	--graphics none \
+	--console pty,target_type=serial \
+	--os-variant ubuntu18.04 \
+	--check all=off \
+	--debug \
+	--qemu-commandline='-sandbox on,obsolete=deny,elevateprivileges=deny,spawn=deny,resourcecontrol=deny'
 
+echo "virt ret=$?"
+sync
 
-qemu-img snapshot -c 'install os' rootfs.qcow2
-qemu-img snapshot -l rootfs.qcow2
+lsof rootfs.qcow2
+if [ $? -ne 0 ]; then
+	size=$(du -s rootfs.qcow2 | awk '{print $1}')
+	if [ "$size" -gt 204800 ]; then
+		qemu-img snapshot -c 'install os' rootfs.qcow2
+		qemu-img snapshot -l rootfs.qcow2
+	else
+		echo "rootfs.qcow2 size too small, skip snapshot"
+	fi
+fi
 
 exit 0
